Update Teleport Point.cs

+++ Enemies/Supplies/Teleport Point.cs
@@ -8,14 +8,14 @@ namespace GwambaPrimeAdventure.Enemy.Supply
 		private
 			ITeleporter _teleporter;
 		private
-			ushort _teleportIndex;
+			byte _teleportIndex;
 		[SerializeField, Tooltip( "If this point will destroy itself after use." ), Header( "Interactions" )]
 		private
 			bool _destroyAfter;
 		[SerializeField, Tooltip( "If this point will trigger with other object." )]
 		private
 			bool _hasTarget;
-		public void GetTouch( ITeleporter teleporter, ushort teleportIndex )
+		public void GetTouch( ITeleporter teleporter, byte teleportIndex )
 		{
 			_teleporter = teleporter;
 			_teleportIndex = teleportIndex;
